Introduce crude reporting of plugin errors
Introduce new error types to help capture relevant information about errors
The intention is to make it easier for users (and contributors), to diagnose
and fix issues if certain plugins fail to initialize.

Ideally we should be able to identify at least:

- Which plugin failed
- Why it failed
- What we can do about it

For a start, simply logging the errors to the console with the nestjs logger
can give us acceptable discoverability of what causes plugins to fail.