feat: modify management of projects to ease iterative workflow
The main purpose of these changes is to allow deleting and recreating
the project without having to setup data sources again.

- Allow deleting a project while keeping its sources
- Create a default project on application start implicitly when none exists
- If leftover sources are found from a previous project, schedule
  indexing for them

This is meant as a shortcut to enable quicker experimentation when
developing an Unbody project. Properly taking care of this requires
solving a few problems that are too complex to tackle right now. This code
is meant to be thrown away once we have more space to do so.